requirements.txt
# Core AWS Dependencies
boto3>=1.34.0
botocore>=1.34.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Optional: ML Explainability
# Uncomment if you want SHAP attribution
# shap>=0.42.0

# Optional: Visualization
# Uncomment for plotting capabilities
# matplotlib>=3.7.0
# seaborn>=0.12.0
